Detect temporary service errors
The reason to be defensive about this case, is that when this error occurs, a
generic page is returned without the feed link, which is expected to
exist on the document. I would'nt want to maybe expect the feed link to
exist on the page since as is the case with most ad-hoc scrapers, it
should be expected that the scraper will break one day, and being fuzzy
about what the page should contain would makes it difficult to fix the
scraper if/when it breaks.